backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0

# Database drivers
psycopg2-binary==2.9.9

# Additional dependencies for user management
email-validator==2.0.0
bcrypt==4.0.1

# File operations and cloud storage
boto3==1.34.131
botocore==1.34.131

# Email services
mailtrap

# PDF generation
reportlab==4.0.4
Pillow==10.0.1

# Production WSGI server
gunicorn==21.2.0
